Test Environments
This change updates the dependency builder to create dependencies for
all of the buildpack's supported environments.  For most dependencies
this is a no-op as they are Java-based and therefore platform
agnostic.  OpenJDK, however, is platform-specific and so it builds for
lucid64, mountainlion64, and precise64.  To achieve this, the OpenJDK
builds are converted to use Vagrant instead of EC2.

[#52860123]