OpenJDK 8 GA Support
This change adds the ability to build OpenJDK 8 GA.  This
functionality should already have been in the builder, but changes to
the OpenJDK build system and source repositories drove the need to
upgrade various bits of the process.  Specifically changes were made
to:

* the version of Mercurial
* the version of the bootstrap JDK
* how the EPEL repository was added

In addition, a change was made to how the AppDynamics agent was
uploaded.  It had been broken by an earlier changed, but was unnoticed
until the latest update.

[#67853180]