Replicate without AWS credentials
As mentioned in #3, the reason for using AWS credentials during
replication is a bit confusing.  Unfortunately, the aws-sdk Ruby gem
doesn't allow anonymous access so this change re-implements the
replication behavior using nothing but REST calls.  This has the side-
effect of speeding up the replication about 30% (by downloading
artifacts from CloudFront) as well.  Finally, the documentation for
replication has been moved to the top of the README file as it is the
most likely to be used by people who are not on the Cloud Foundry Java
Experience team.

[resolves #3]