Trusty Java 7 Builds
This change enables the builder to create Trusty Java 7 builds.  
Previously the builder would fail trying to build them because the
location of the installed Java 6 JDK had changed.  There is now a
special case for Trusty.

In addition, this change puts in a fix for detecting the number of cores
on ubuntu machines.  I believe that this detection has been broken since
day 1. leading to single core (and therefore incredibly slow) builds for
ubuntu OSes.