Use full path when downloading
Previously, the pump utility only used the "path" of a URI when
downloading.  This path did not include the query parameters that
might be attached to the URI.  This meant that the latest Groovy
version would not build properly.  This change causes the pump to use
request_uri which is the path, plus the query parameters.