Testing of Node and Ruby Builds
Previously, Node and Ruby builds were created quickly in order to get
the artifacts that they produces built to the repository.  This
necessitated skipping over a lot of the testing and refactoring that
should have been done.  This change adds that testing and refactoring
so that the code-base is of the proper level of quality.