Vagrant Rebuild
Previously there were some performance problems with the Vagrant
configuration that led to excessively long build times for certain
components.  This change rebuilds the Vagrant configuration to
eliminate these performance problems.

Note that this also has the start of an attempt to get Mountain Lion
into Vagrant, but it is not yet complete.