Use SSL when replicating
Previously, the replication script always assumed that the downloads
would happen from a non-SSL endpoint.  With the recent addition of
SSL, this was no longer true, but the replication script did not start
using SSL at this time which caused 400 failures to be returned.  This
change updates the replication script to always use SSL to download
artifacts as this is (currently) the most secure way to download the
artifacts.