Remove AppDynamics Latest Behavior
Previously, the latest version of the AppDynamics dependency couldn't
be retrieved via a URL with a version number in it.  This lead to the
"latest" version of  AppDynamics being treated differently.  Recently
the downloads have changed and the latest version is available via a
URL with a version number in it.  This change removes the "latest"
functionality.