Normalize Memory Calculator Versions
Previously, the versions of the memory calculator were passed through
without change.  This meant that they stayed the SpringSource standard
using dots between all components.  This version scheme does not work
with the Java Buildpack and broke the version resolution of the memory
calculator.

This change adds in the MMMQ normalization to ensure that version
resolution works.