Harmonize Resources
Previously, each of the Concourse resources was effectively a standalone
project that copied-and-pasted a lot of code from the others.  This change
harmonizes those including shared code, and common base classes enforcing
behavior.  This should make things quite a bit more maintainable as well as
creating a "support" library for future Java-based resources.