Remove Non-Open Source from Replication
Previously, all of the dependencies required for the Java Buildpack were
downloaded as part of the replication script.  Given recent guidance from the
Cloud Foundry Foundation, dependencies that are not backed by pure open source
can be integrated but should not be available from CFF domains.  This is a
step towards that guidance removing non-open source dependencies from the
replications script.